requirements.txt
flask-fs-router
flask
click
questionary
python-dotenv
requests
googlesearch-python
searchit
free-proxy
pydantic
craiyon.py
pillow
dnspython
docker
colorama
yaspin
cronitor
pyngrok
ollama
